server/requirements.txt
Django>=5.0
djangorestframework>=3.14
drf-spectacular>=0.27
celery>=5.3
redis>=5.0
python-dotenv>=1.0
django-cors-headers>=4.3
pytest>=8.0
pytest-django>=4.7
pytest-mock>=3.0
openai>=1.0
jsonschema>=4.0
